requirements.txt
aiohttp
psycopg2-binary
python-telegram-bot[job-queue]
python-dotenv
requests
flask